moved filters into their own reducer
