+++ src/components/Visualizations/Contact311.jsx
@@ -85,7 +85,7 @@ const Contact311 = () => {
 };
 
 const mapStateToProps = state => ({
-  requestTypes: state.data.requestTypes,
+  requestTypes: state.filters.requestTypes,
 });
 
 export default connect(mapStateToProps)(Contact311);
