+++ src/components/Visualizations/Frequency.jsx
@@ -82,7 +82,7 @@ const Frequency = ({
 };
 
 const mapStateToProps = state => ({
-  requestTypes: state.data.requestTypes,
+  requestTypes: state.filters.requestTypes,
 });
 
 export default connect(mapStateToProps)(Frequency);
