+++ src/components/Visualizations/Legend.jsx
@@ -42,7 +42,7 @@ const Legend = ({
 };
 
 const mapStateToProps = state => ({
-  requestTypes: state.data.requestTypes,
+  requestTypes: state.filters.requestTypes,
 });
 
 export default connect(mapStateToProps)(Legend);
