+++ src/components/Visualizations/TotalRequests.jsx
@@ -87,7 +87,7 @@ const TotalRequests = ({
 };
 
 const mapStateToProps = state => ({
-  requestTypes: state.data.requestTypes,
+  requestTypes: state.filters.requestTypes,
 });
 
 export default connect(mapStateToProps)(TotalRequests);
