+++ src/components/Visualizations/TypeOfRequest.jsx
@@ -62,7 +62,7 @@ const TypeOfRequest = ({
 };
 
 const mapStateToProps = state => ({
-  requestTypes: state.data.requestTypes,
+  requestTypes: state.filters.requestTypes,
 });
 
 export default connect(mapStateToProps)(TypeOfRequest);
