+++ src/redux/rootReducer.js
@@ -1,8 +1,10 @@
 import { combineReducers } from 'redux';
 import data from './reducers/data';
+import filters from './reducers/filters';
 import ui from './reducers/ui';
 
 export default combineReducers({
   data,
+  filters,
   ui,
 });
